Revert "VertexDrawRange Tweaks"
This reverts commit f0852fc3d92f86e3bc17941c56138048e79ddb21.

+++ Source/Core/Duality/Drawing/VertexData/VertexDrawRange.cs
@@ -17,11 +17,11 @@ namespace Duality.Drawing
 		/// <summary>
 		/// Index of the first vertex to be rendered.
 		/// </summary>
-		public ushort Index;
+		public int Index;
 		/// <summary>
 		/// The number of vertices to be rendered, starting from <see cref="Index"/>.
 		/// </summary>
-		public ushort Count;
+		public int Count;
 
 		public override string ToString()
 		{
